pyproject.toml
[project]
name = "llm-agent-framework"
version = "1.0.0"
description = "Generic, config-driven LLM agent framework"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "anthropic>=0.40.0",
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.6",
    "alembic>=1.14.0",
    "structlog>=24.4.0",
    "langfuse>=2.53.0",
    "sentry-sdk[fastapi]>=2.19.0",
    "httpx>=0.28.0",
    "sse-starlette>=2.2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "httpx>=0.28.0",
]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["B008"]  # FastAPI uses Depends() in function defaults by design

[tool.ruff.format]
quote-style = "single"

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src", "agents"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
